% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trackman_functions.R
\name{get_transverse_spin}
\alias{get_transverse_spin}
\title{Get Transverse Spin}
\usage{
get_transverse_spin(extension, tot_spin, vx0, vy0, vz0, ax, ay, az,
  direction = "TRANSVERSE")
}
\arguments{
\item{extension}{Release Extension (In Feet From Mound)}

\item{tot_spin}{Release Total Spin Rate (RpM)}

\item{vx0}{Initial X Velocity (MpH)}

\item{vy0}{Initial Y Velocity (MpH)}

\item{vz0}{Initial Z Velocity (MpH)}

\item{ax}{X Acceleration}

\item{ay}{Y Acceleration}

\item{az}{Z Acceleration}

\item{direction}{Which spin component should be returned? See details}
}
\value{
If directional = TRUE, the function will return a named list with
total transverse spin as well as the spin in the X,Y, and Z directions,
otherwise it will return a scalar representing the total transverse spin.
}
\description{
Get Transverse Spin takes the Trackman data that is publicly available and
uses it to calculate the trasverse (useful) spin. The calculations in this
package are based on the following paper by Dr. Allen Nathan:
http://baseball.physics.illinois.edu/trackman/SpinAxis.pdf
and all calculations were directly adapted from this workbook, also by
Dr. Nathan:
http://baseball.physics.illinois.edu/trackman/MovementSpinEfficiencyTemplate.xlsx
}
\details{
The direction argument will return the total transverse spin if
direction=="TRANSVERSE". "X","Y", or "Z" will return the transvers spin in
that direction. Direction == "ALL" will return a named list consisting of
the transverse spin (spinT) as well as all components (spinTx,spinTy,spinTz)
}
\note{
The current version of this function is based on atmospheric
conditions inside Tropicana Field in June, but future versions should
include paramters to specify atmospheric condiditions so as to allow for
more accurate spin estimates across ballparks.
}
